moved token and session stuff to own folder
