+++ app/src/main/java/com/cornellappdev/uplift/ui/viewmodels/nav/RootNavigationViewModel.kt
@@ -1,7 +1,7 @@
 package com.cornellappdev.uplift.ui.viewmodels.nav
 
 import androidx.lifecycle.viewModelScope
-import com.cornellappdev.uplift.data.repositories.SessionManager
+import com.cornellappdev.uplift.data.auth.SessionManager
 import com.cornellappdev.uplift.data.repositories.UserInfoRepository
 import com.cornellappdev.uplift.ui.UpliftRootRoute
 import com.cornellappdev.uplift.ui.nav.RootNavigationRepository
